+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -59,6 +59,4 @@
 	<mybatis-spring:scan base-package="com.coderby.myapp.attendance.dao"/>
 	<context:component-scan base-package="com.coderby.myapp.attendance.service"/>
 	
-	
-	
 </beans>
